test(check_arg): move test cases to public api folder
