+++ tests/src/test_cases/public_api/test_check_obj_impl.h
