+++ tests/src/test_cases/public_api/test_check_obj_no_cls_no_vld.c
@@ -1,8 +1,9 @@
 #if LV_BUILD_TEST
-/* Variant: LV_USE_CHECK_ARG=1, LV_USE_CHECK_OBJ_CLASSTYPE=1, LV_USE_CHECK_OBJ_VALIDITY=0 */
+/* Variant: LV_USE_CHECK_ARG=1, LV_USE_CHECK_OBJ_CLASSTYPE=0, LV_USE_CHECK_OBJ_VALIDITY=0 */
+#define LV_USE_CHECK_OBJ_CLASSTYPE 0
 #define LV_USE_CHECK_OBJ_VALIDITY 0
 #include "../lvgl.h"
-#include "test_cases/test_check_obj_impl.h"
+#include "test_cases/public_api/test_check_obj_impl.h"
 
 void setUp(void)
 {
