+++ tests/src/test_cases/public_api/test_obj_event_invariants.c
